fix: thumbnail view styles
