+++ frontend/src/components/NavigationPanel.vue
@@ -22,27 +22,7 @@
 				@click="handleSlideClick(slide)"
 				:ref="(el) => (slideThumbnailsRef[sidebarSlidesList.indexOf(slide)] = el)"
 			>
-				<div :style="getThumbnailInnerStyles(slide)">
-					<SlideElement
-						v-for="element in slide?.elements"
-						:key="`export-${element.id}`"
-						mode="export"
-						:element="element"
-						:data-index="element.id"
-					/>
-				</div>
-				<div
-					class="absolute inset-0 flex justify-between rounded p-2"
-					:style="getGradientOverlayStyles(slide)"
-				>
-					<div class="text-[10px] font-medium">{{ slide.idx }}</div>
-					<TransitionIcon v-if="slide.transition != 'None'" class="h-3 opacity-80" />
-				</div>
-
-				<div
-					v-if="isSlideActive(slide)"
-					class="absolute -left-5 h-full w-2 rounded-r bg-blue-500 opacity-90"
-				></div>
+				<ThumbnailContainer :slide="slide" :isActive="isSlideActive(slide)" />
 			</div>
 		</div>
 	</div>
@@ -58,14 +38,11 @@ import { ref, computed, watch, nextTick, useTemplateRef, useAttrs, inject } from
 
 import Draggable from 'vuedraggable'
 
-import TransitionIcon from '@/icons/TransitionIcon.vue'
-import SlideElement from '@/components/SlideElement.vue'
-
+import ThumbnailContainer from '@/components/ThumbnailContainer.vue'
 import { useNavigationPanel } from '@/composables/useNavigationPanel'
 
 import { slides, slideIndex, currentSlide, focusedSlide } from '@/stores/slide'
 import { handleScrollBarWheelEvent, getThumbnailCardStyles } from '@/utils/helpers'
-import { isBackgroundColorDark } from '@/utils/color'
 
 import { commandHistory, recentlyRestored } from '@/stores/historyMeta'
 import { reorderSlidesCommand } from '@/stores/commands'
@@ -78,19 +55,6 @@ const scrollableArea = useTemplateRef('scrollableArea')
 const { isNavigationPanelOpen, toggleNavigationPanel } = useNavigationPanel()
 
 const inReadonlyMode = inject('inReadonlyMode', ref(false))
-
-const getGradientOverlayStyles = (slide) => {
-	const hasDarkBg = isBackgroundColorDark(slide.background)
-	const textColor = hasDarkBg ? '#ffffff' : '#00000090'
-	const background = hasDarkBg
-		? 'linear-gradient(140deg, rgba(255, 255, 255, 0.3) 0%, rgba(255, 255, 255, 0) 20%, rgba(0, 0, 0, 0) 100%)'
-		: 'linear-gradient(140deg, rgba(0, 0, 0, 0.1) 0%, rgba(0, 0, 0, 0) 20%, rgba(0, 0, 0, 0) 100%)'
-
-	return {
-		background,
-		color: textColor,
-	}
-}
 
 const emit = defineEmits(['changeSlide', 'openLayoutDialog'])
 
@@ -157,16 +121,6 @@ const getThumbnailStyles = (s) => {
 		height: `${540 * THUMBNAIL_SCALE}px`,
 	}
 }
-
-const getThumbnailInnerStyles = () => ({
-	width: '960px',
-	height: '540px',
-	transformOrigin: 'top left',
-	transform: `scale(${THUMBNAIL_SCALE})`,
-	position: 'absolute',
-	top: '0',
-	left: '0',
-})
 
 const toggleButtonClasses = computed(() => {
 	const baseClasses = 'flex cursor-pointer items-center border bg-white'
